frontend_flask/requirements.txt
flask
flask_sqlalchemy
psycopg2-binary
python-dotenv
requests
werkzeug
pytest